Add missing type files
