+++ package.json
@@ -43,6 +43,7 @@
     "@types/node": "^20.12.7",
     "@types/nodemailer": "^6.4.11",
     "@types/react": "19.1.10",
+    "@types/react-dom": "^19.1.9",
     "@types/uuid": "^9.0.7",
     "prisma": "^6.14.0",
     "tsx": "^4.19.2",
