+++ src/app/professional/settings/page.tsx
@@ -1,7 +1,7 @@
 import { Card, Button, Input } from "../../../components/ui";
 import { auth } from "../../../../auth";
 import { getProfessionalSettings } from "../../api/professional/settings";
-import { useState, useEffect } from "react";
+import StripeSection from "./StripeSection";
 
 export default async function ProSettings() {
   const session = await auth();
@@ -37,77 +37,6 @@ export default async function ProSettings() {
         </div>
       </div>
     </Card>
-  );
-}
-
-function StripeSection() {
-  'use client';
-  const [status, setStatus] = useState('loading');
-  const [loading, setLoading] = useState(false);
-  const [message, setMessage] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    async function fetchStatus() {
-      try {
-        const res = await fetch('/api/stripe/account');
-        if (!res.ok) throw new Error();
-        const data = await res.json();
-        setStatus(data.status || 'not_connected');
-      } catch {
-        setError('Failed to load Stripe status');
-      }
-    }
-    fetchStatus();
-  }, []);
-
-  const handleConnect = async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const res = await fetch('/api/stripe/account', { method: 'POST' });
-      const data = await res.json();
-      if (!res.ok || !data.onboardingUrl) throw new Error();
-      setMessage('Redirecting to Stripe...');
-      window.location.href = data.onboardingUrl;
-    } catch {
-      setError('Failed to start Stripe onboarding');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const statusLabel =
-    status === 'complete'
-      ? 'Connected'
-      : status === 'pending'
-      ? 'Pending verification'
-      : status === 'incomplete'
-      ? 'Incomplete'
-      : status === 'loading'
-      ? 'Loading...'
-      : 'Not connected';
-
-  const canConnect =
-    status === 'not_connected' || status === 'incomplete';
-
-  return (
-    <>
-      <h3>Payment</h3>
-      <p>Onboard to Stripe Connect to receive payouts.</p>
-      {canConnect && (
-        <Button onClick={handleConnect} disabled={loading}>
-          {loading
-            ? 'Loading...'
-            : status === 'incomplete'
-            ? 'Resume Onboarding'
-            : 'Connect Stripe'}
-        </Button>
-      )}
-      <p>Account status: {statusLabel}</p>
-      {error && <p style={{ color: 'red' }}>{error}</p>}
-      {message && <p style={{ color: 'green' }}>{message}</p>}
-    </>
   );
 }
